FlowerRecognition/requirements-extra.txt
python-multipart
fastapi
uvicorn
pillow